length

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -3,9 +3,10 @@
  * @s: string
  * Return: length 
  **/
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
 	size_t len = 0;
+
 	while ( *str++)
 		len++;
 	return (len);
